requirements.txt
django
djangorestframework
requests
django-crispy-forms